Refactor kontellation to constellation + change packages name
