--- src/main/kotlin/org/web3j/qourum/konstellation/http/Method.kt
@@ -1,4 +1,0 @@
-package io.blk.konstellation.http
-
-const val GET = "GET"
-const val POST = "POST"
