+++ src/main/kotlin/org/web3j/quorum/enclave/Types.kt
@@ -1,5 +1,4 @@
-package io.blk.konstellation
-
+package org.web3j.quorum.enclave
 
 class SendRequest(
         val payload: String,
