+++ src/main/kotlin/org/web3j/quorum/enclave/http/Method.kt
@@ -1,0 +1,4 @@
+package org.web3j.quorum.enclave.http
+
+const val GET = "GET"
+const val POST = "POST"
