+++ src/main/kotlin/org/web3j/quorum/enclave/http/Response.kt
@@ -1,4 +1,5 @@
-package io.blk.konstellation.http
+package org.web3j.quorum.enclave.http
+
 
 data class Response(
         val statusCode: StatusCode,
