+++ src/main/kotlin/org/web3j/quorum/enclave/http/ResponseParserException.kt
@@ -1,4 +1,4 @@
-package io.blk.konstellation.http
+package org.web3j.quorum.enclave.http
 
 class ResponseParserException : RuntimeException {
     constructor(message: String): super(message) {}
