+++ src/test/java/org/web3j/quorum/RawTransactionManagerTest.java
@@ -13,7 +13,7 @@ import org.junit.Test;
 import org.web3j.crypto.Credentials;
 import org.web3j.crypto.WalletUtils;
 import org.web3j.protocol.http.HttpService;
-import org.web3j.qourum.tx.QuorumTransactionManager;
+import org.web3j.quorum.enclave.tx.QuorumTransactionManager;
 import org.web3j.quorum.generated.Greeter;
 import org.web3j.quorum.tx.ClientTransactionManager;
 import org.web3j.tx.Contract;
@@ -22,8 +22,8 @@ import org.web3j.tx.gas.DefaultGasProvider;
 import static junit.framework.TestCase.assertTrue;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
-import static org.web3j.qourum.util.Base64Kt.decode;
-import static org.web3j.qourum.util.Base64Kt.encode;
+import static org.web3j.quorum.enclave.util.Base64Kt.decode;
+import static org.web3j.quorum.enclave.util.Base64Kt.encode;
 import static org.web3j.tx.Contract.GAS_LIMIT;
 import static org.web3j.tx.ManagedTransaction.GAS_PRICE;
