+++ src/test/kotlin/org/web3j/quorum/enclave/http/RequestBuilderTest.kt
@@ -1,4 +1,4 @@
-package io.blk.konstellation.http
+package org.web3j.quorum.enclave.http
 
 import org.assertj.core.api.Assertions.*
 import org.junit.Test
@@ -15,7 +15,7 @@ class RequestBuilderTest {
                 """
                 $POST /$path HTTP/1.1
                 Host: k
-                User-Agent: konstellation-client
+                User-Agent: constellation-client
                 Content-Type: application/json
                 Content-Length: ${json.length}
